LibJS: Treat the ISO 8601 calendar the same as the Gregorian calendar
This matches other engines, and further allows our Intl.DateTimeFormat
implementation to satisfy the general requirement that users may use any
calendar returned by `Intl.supportedValuesOf("calendar")`. This lets us
replace the NotImplemented exception in ToLocalTime with a TODO(), as
it now cannot be reached (until we implement more calendars).

An upcoming commit will mark some calls to FormatDateTime (and thus to
ToLocalTime) as infallible. This patch allows that change, as callers
will not need to worry about internal exceptions.